修改cmd_set_normal_mode为cmd_set_work_mode
