--- code/src/WechatAPI/TestModeShell/normal_mode_config.txt
@@ -1,3 +1,0 @@
-1AAA02020045
-1AAA02020046
-1AAA02020047
