+++ code/src/WechatAPI/TestModeShell/work_mode_config.txt
@@ -1,0 +1,3 @@
+1AAA02020045
+1AAA02020046
+1AAA02020047
